refactor(func): move chain() to own file; drop unused MoreLinq usings
- Move chain() out of ListFunctions.cs into ListFunctions.Chain.cs so it matches
  its sibling pipeline functions (each in its own ListFunctions.<Name>.cs), per
  review. No behavior change; ListFunctionUnitTests green (290).
- Remove the unused `using MoreLinq.Extensions;` from the EverySome/FilterQ/
  JsonGroupBy/Jiter partials (leftover from copying the ListFunctions.cs using
  block verbatim). Build stays green with TreatWarningsAsErrors.
